Transferred the routes I added in my own example as well as required types. "connectedRoute" (through "/wheretogo") sends a JSON list of edges connected to the current vertex. "getGraphRoute" (through "/getGraph") returns the same, but also includes the current vertex. For the latter, edges that connect to the current vertex, not just outbound edges, are included.
This does not include the functionality I added in my example to load the TF metadata and export that to JSON during a move or whereami action; I can add that if you want.

ExampleServer contains a small node server that produces a visualisation of the network using vis.js. It uses the "getGraph" route to retrive the current vertex and edges and then draw it as a network. Clicking on a node will send a request to the move route, and recreate the visualisation (through a page reload) based on the next zipper location. Clicking the Back button will attempt to use the back route to return to the previous vertex. Both inbound (in red) and outbound (in green) edges are included, though only the latter can be used to navigate right now.

The node_modules folder is not included with the server so run "npm install" first before using "npm start" to run the server. It is accessible at "localhost:3000".

Signed-off-by: n7581769 <st2.smith@hdr.qut.edu.au>